+++ LVL5/webserv/utils2.cpp
@@ -6,14 +6,14 @@
 /*   By: hdrabi <hdrabi@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/26 13:48:33 by hdrabi            #+#    #+#             */
-/*   Updated: 2022/09/26 18:10:43 by hdrabi           ###   ########.fr       */
+/*   Updated: 2022/09/27 12:08:25 by hdrabi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "utils2.hpp"
 
 int get_port(std::string listen) {
-	int i = 0;
+	size_t i = 0;
 	std::string rst = listen;
 
 	i = listen.find(':');
@@ -21,7 +21,7 @@ int get_port(std::string listen) {
 		rst = listen.substr(i + 1, listen.length());
 	if (rst.empty())
 		throw std::runtime_error("configFile error, please check your port");
-	for (int j = 0; j < rst.length() ; j++) {
+	for (size_t j = 0; j < rst.length() ; j++) {
 		if (!std::isdigit(rst[j])) {
 			if (i != std::string::npos)
 				throw std::runtime_error("configFile error, wrong port format");
@@ -36,13 +36,13 @@ int get_port(std::string listen) {
 }
 
 std::string get_host(std::string listen) {
-	int i = 0;
+	size_t i = 0;
 	std::string rst = listen;
 
 	i = listen.find(':');
 	if (i != std::string::npos)
 		rst = listen.substr(0, i);
-	for (int j = 0; j < rst.length() ; j++) {
+	for (size_t j = 0; j < rst.length() ; j++) {
 		if (!std::isdigit(rst[j])) {
 				return rst;
 		}
@@ -52,7 +52,7 @@ std::string get_host(std::string listen) {
 }
 
 std::vector<std::string> split_vec(std::string str) {
-	int i = 0;
+	size_t i = 0;
 	std::vector<std::string> vec;
 	std::string rst;
